modified the function
modified function to ask the user before signing out

+++ SigningOut.py
@@ -8,9 +8,15 @@ class SigningOut(QMainWindow):
     def signOut(self):
         print("out")
         if os.path.exists('token.pickle'):
-            os.remove("token.pickle")
-            QMessageBox.question(self, 'Message', "Now you are signd out", QMessageBox.Ok,
-                            QMessageBox.Ok)
+            buttonReply = QMessageBox.question(self, 'Message', "Do you do u want to sign out ?",
+                                                           QMessageBox.Yes | QMessageBox.No, QMessageBox.No)
+            if buttonReply == QMessageBox.Yes:
+                os.remove("token.pickle")
+                QMessageBox.question(self, 'Message', "Now you are signd out", QMessageBox.Ok,
+                                QMessageBox.Ok)
+            return 1
+
         else:
             QMessageBox.question(self, 'Message', "you are not connected to any google drive", QMessageBox.Ok,
                                  QMessageBox.Ok)
+            return 0
